requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
black==23.10.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
requests==2.31.0
slack-bolt==1.18.0
slack-sdk==3.23.0
tinydb==4.8.0
urllib3==2.0.7
yarl==1.9.2
python-dotenv==1.0.0
pynamodb==5.5.0